path
Provides animation along bezier and circular arcs.

The animation engine in jQuery is focussed on single dimensional
animation - hence it's difficult to animate two variables along a path.

This simple plugin provides a method of multidimensional animation, and
in particular provides a method for animating along bezier curves and
arcs.